Check if the path exists before chmod operation in Create Mavericks Installer.tool
In macOS 10.12 beta 2, the Finder will not create a .Trashes directory
automatically.
The chmod operation on a non existing file will cause the script return
failed.
And the install process will be terminated and reported as failed.

The solution is to check if the path exists before the chmod operation.